html/browsers/history/the-location-interface/location-replace-from-iframe.sub.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>Referer and location.replace</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <iframe src="/resources/blank.html" hidden></iframe>
    <script>
      async_test(function(t) {
        function on_message(e) {
        const referrer = e.data;
        assert_equals(referrer, window.location.href);
        t.done();
      }
      window.addEventListener('message', t.step_func(on_message), false);
      document.querySelector("iframe").contentWindow.location.replace("resources/iframe-contents.sub.html");
      }, "Browser sends Referer header in iframe request when location.replace is called from an iframe");
    </script>
  </body>
</html>
